+++ db/db_test.cc
@@ -1349,7 +1349,6 @@ TEST_F(DBTest, Snapshot) {
 
 TEST_F(DBTest, HiddenValuesAreRemoved) {
   do {
-    std::cout<<"In this test"<<std::endl;
     Random rnd(301);
     FillLevels("a", "z");
