+++ mod/learned_merger.cc
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "table/learned_merger.h"
-
 #include "leveldb/comparator.h"
 #include "leveldb/iterator.h"
 #include "table/iterator_wrapper.h"
+#include "mod/learned_merger.h"
 #include "mod/plr.h"
+
 #include <iostream>
 #include <cmath>
